backend/requirements.txt
django
djangorestframework
django-cors-headers
numpy
scipy
scikit-learn
matplotlib
pandas